feat(readme): added my personal technology stack
This stack is intended to show which tools I prefer to work with and have gained practical experience with. It by no means covers my entire skillset and much of this knowledge can be transferred to other technologies as well.

Unfortunately, I also noticed that some symbols are really hard to recognize in dark mode. Also, some icons, such as the one for GitLab, have recently become outdated.